+++ fluent_pages/models/db.py
@@ -57,7 +57,6 @@ class UrlNode(with_metaclass(URLNodeMetaClass, PolymorphicMPTTModel, Translatabl
     """
     The base class for all nodes; a mapping of an URL to content (e.g. a HTML page, text file, blog, etc..)
     """
-
     # Some publication states
     DRAFT = 'd'
     PUBLISHED = 'p'
@@ -77,6 +76,7 @@ class UrlNode(with_metaclass(URLNodeMetaClass, PolymorphicMPTTModel, Translatabl
     publication_date = models.DateTimeField(_('publication date'), null=True, blank=True, db_index=True, help_text=_('''When the page should go live, status must be "Published".'''))
     publication_end_date = models.DateTimeField(_('publication end date'), null=True, blank=True, db_index=True)
     in_navigation = models.BooleanField(_('show in navigation'), default=appsettings.FLUENT_PAGES_DEFAULT_IN_NAVIGATION, db_index=True)
+    in_sitemaps = models.BooleanField(_('include in search engine sitemaps'), default=True, db_index=True)
     override_url = TranslatedField()
 
     # For tagging nodes and locating them in code. This should be avoided if possible,