+++ fluent_pages/sitemaps.py
@@ -27,7 +27,7 @@ class PageSitemap(Sitemap):
         Return all items of the sitemap.
         """
         # Note that .active_translations() can't be combined with other filters for translations__.. fields.
-        return UrlNode.objects.published().non_polymorphic().active_translations() \
+        return UrlNode.objects.in_sitemaps().non_polymorphic().active_translations() \
                 .order_by('level', 'translations__language_code', 'translations___cached_url')
 
     def lastmod(self, urlnode):