+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changes in version 0.9 (dev)
 * Created ``fluent_pages.integration.fluent_contents`` to simplify creating custom
   pagetypes that reuse the django-fluent-contents_ integration that the ``fluent_pages.pagetypes.fluentpage`` has.
 * Added ``CurrentPageMixin`` and ``CurrentPageTemplateMixin`` for custom views.
+* Added ``in_sitemaps`` flag, which is now false for the ``RedirectNode`` by default.
 * API: use ``FluentContentsPage`` instead of ``AbstractFluentPage``.